+++ src/recallzero/recall/__init__.py
@@ -1,3 +1,9 @@
 from recallzero.recall.matcher import RecallMatcher, normalize_match_text
+from recallzero.recall.target_attributor import TargetAttributionScore, TargetAttributor
 
-__all__ = ["RecallMatcher", "normalize_match_text"]
+__all__ = [
+    "RecallMatcher",
+    "TargetAttributionScore",
+    "TargetAttributor",
+    "normalize_match_text",
+]